File Refactorize and Test easy env
