+++ ref/test/util_test.c
@@ -4,437 +4,41 @@ struct timeval start, end;
 double time_used;
 clock_t c_start, c_end;
 
-int main(){
-    FILE *fp = NULL;
-    double op_total_time[12] = {0, };
-    double op_exp_time[3] = {0, };
-    byte temp[1] = {0};
-    int test_word_size_a = test_word_size;
-    int test_word_size_b = test_word_size;
-    int test_word_size_c = test_word_size;
-    int test_max_word_size = test_word_size;
-    int karachuba_flag = 0;
-    int squ_karachuba_flag = 0;
-    int DIVISION_METHOD = 0;
-    int miller_rabin_iter = 10;
+msg util_test(){
+    if(test_bi_new_delete() != Test_BI_NEW_DELETE_SUCCESS)    return Test_FAIL;
+    if(test_bi_random() != Test_BI_GET_RANDOM_SUCCESS)    return Test_FAIL;
+    if(test_bi_set_from() != Test_BI_SET_FROM_SUCCESS)    return Test_FAIL;
+    if(test_bi_shift() != Test_BI_SHIFT_SUCCESS)    return Test_FAIL;
+    if(test_bi_get_lower() != Test_BI_GET_LOWER_SUCCESS)    return Test_FAIL;
+    if(test_bi_cat() != Test_BI_CAT_SUCCESS)    return Test_FAIL;
+    return Test_SUCCESS;
+}
+
+msg new_bigint_test_vector(){
+    msg result_msg = new_bigint_test_vector_FAIL;
     char* str = NULL;
-    msg result_msg = Test_SUCCESS;
     bigint* a = NULL;
-    bigint* b = NULL;
-    bigint* c = NULL;
-    bigint* barret_N = NULL;
-    bigint* barret_T = NULL;
 
-    // 테스트 빅인트 파일 초기화
-//    Test_file_write_non_enter(Test_bigint, "", CLEAR);
-    Test_file_write_non_enter(Test_file_add, "", CLEAR);
-    Test_file_write_non_enter(Test_file_sub, "", CLEAR);
-    Test_file_write_non_enter(Test_file_mul, "", CLEAR);
-    Test_file_write_non_enter(Test_file_mul_karachuba, "", CLEAR);
-    Test_file_write_non_enter(Test_file_div, "", CLEAR);
-    Test_file_write_non_enter(Test_file_squ_karachuba, "", CLEAR);
-    Test_file_write_non_enter(Test_file_exp, "", CLEAR);
-    Test_file_write_non_enter(Test_file_barrett_reduction, "", CLEAR);
-    Test_file_write_non_enter(Test_file_gcd, "", CLEAR);
-    Test_file_write_non_enter(Test_file_EEA, "", CLEAR);
-    Test_file_write_non_enter(Test_file_MR, "", CLEAR);
-    Test_file_write_non_enter(Test_file_squ, "", CLEAR);
+    // file clear
+    if(Test_file_write_non_enter(Test_bigint, "", CLEAR) != FILE_WRITE_SUCCESS)   return new_bigint_test_vector_FAIL;
 
-//    clock_t start = 0;
-//    clock_t end = 0;
-//    for(int i = 0; i < 1; i++){
-//        start = clock();
-////        result_msg = RSA();
-//        result_msg = RSA_CRT();
-//        end = clock();
-//        log_msg(result_msg);
-//    }
-//    printf("time : %f\n", (double)(end - start) / CLOCKS_PER_SEC);
-//    return 0;
-
-
-//    test_bi_new_delete();       // bigint 할당 및 해제 테스트
-//    test_bi_random();           // 랜덤 bigint 생성 테스트
-//    test_bi_set_from();         // string으로부터 bigint 생성 테스트
-//    test_bi_shift();            // bigint shift 테스트
-//    test_bi_get_lower();        // bigint modular 테스트
-//    test_bi_cat();              // bigint cat 테스트
-
-    // barret N 생성
-    result_msg = init_barret_N(&barret_T, &barret_N, barret_word_size);
-    if (result_msg != INIT_BARRET_N_SUCCESS){
-        log_msg(result_msg);
-        return Test_FAIL;
-    }
-
-    // 카라츄바 세팅
-    if(init_karachuba_pool(test_word_size) != INIT_KARACHUBA_POOL_SUCCESS){
-        log_msg(INIT_KARACHUBA_POOL_FAIL);
-        return Test_FAIL;
-    }
+    str = (char *)calloc((test_word_size * 8) * 4 + 100, sizeof(char));
+    if (str == NULL)    return MEM_NOT_ALLOC;
 
     for(int i = 0; i < test_size; i++){
-        // test_word_size가 0보다 작거나 같으면 랜덤으로 test_word_size를 할당
-        if(test_word_size <= 0){
-            do{
-                if(randombytes(temp, 1) != GEN_RANDOM_BYTES_SUCCESS)    return GEN_RANDOM_BYTES_FAIL;
-                test_word_size_a = temp[0] % test_word_size_limit;
-            }while(test_word_size_a <= 0);
-            do{
-                if(randombytes(temp, 1) != GEN_RANDOM_BYTES_SUCCESS)    return GEN_RANDOM_BYTES_FAIL;
-                test_word_size_b = temp[0] % test_word_size_limit;
-            }while(test_word_size_b <= 0);
-            do{
-                if(randombytes(temp, 1) != GEN_RANDOM_BYTES_SUCCESS)    return GEN_RANDOM_BYTES_FAIL;
-                test_word_size_c = temp[0] % test_word_size_limit;
-            }while(test_word_size_c <= 0);
-            test_max_word_size = max(test_word_size_a, test_word_size_b);
-            test_max_word_size = max(test_max_word_size, test_word_size_c);
+        for(int j = 0; j < 3; j++){
+            result_msg = get_random_bigint(1, &a);
+            if(result_msg != BI_GET_RANDOM_SUCCESS)    goto NEW_BIGINT_EXIT;
+            bigint_to_hex(str, &a);
+            Test_file_write_non_enter(Test_bigint, str, APPEND);
+            Test_file_write_non_enter(Test_bigint, "\n", APPEND);
         }
-
-        // 랜덤으로 a, b, c 생성
-        result_msg = bi_get_random(&a, test_word_size_a);
-        if (result_msg != BI_GET_RANDOM_SUCCESS)    goto TEST_EXIT;
-        else if(a->word_len != test_word_size_a){
-            result_msg = BI_GET_RANDOM_LENGTH_NOT_MATCH;
-            goto TEST_EXIT;
-        }
-
-        result_msg = bi_get_random(&b, test_word_size_b);
-        if (result_msg != BI_GET_RANDOM_SUCCESS)    goto TEST_EXIT;
-        else if(b->word_len != test_word_size_b){
-            result_msg = BI_GET_RANDOM_LENGTH_NOT_MATCH;
-            goto TEST_EXIT;
-        }
-
-        result_msg = bi_get_random(&c, test_word_size_c);
-        if (result_msg != BI_GET_RANDOM_SUCCESS)    goto TEST_EXIT;
-        else if(b->word_len != test_word_size_b){
-            result_msg = BI_GET_RANDOM_LENGTH_NOT_MATCH;
-            goto TEST_EXIT;
-        }
-
-        // 문자열 담을 배열 선언 크기는 다시 고려
-        str = (char *)calloc((test_max_word_size * 8) * 4 + 100, sizeof(char)); // '0x' * 3 + 부호 * 3 + " + " + " = " => 6 + 3 + 3 + 3 = 15
-        if (str == NULL){
-            result_msg = MEM_NOT_ALLOC;
-            goto TEST_EXIT;
-        }
-
-//        // bigint 생성
-//        bigint_to_hex(str, &a);
-//        Test_file_write_non_enter(Test_bigint, str, APPEND);
-//        Test_file_write_non_enter(Test_bigint, "\n", APPEND);
-//
-//
-//        bigint_to_hex(str, &b);
-//        Test_file_write_non_enter(Test_bigint, str, APPEND);
-//        Test_file_write_non_enter(Test_bigint, "\n", APPEND);
-//
-//
-//        bigint_to_hex(str, &c);
-//        Test_file_write_non_enter(Test_bigint, str, APPEND);
-//        Test_file_write_non_enter(Test_bigint, "\n", APPEND);
-/*
-       // bigint 덧셈 테스트
-//       result_msg = test_bi_add(&op_total_time[0], &nums[0], &nums[1], str);
-       result_msg = test_bi_add(&op_total_time[0], &a, &b, str);
-       if(result_msg != Test_BI_ADD_SUCCESS){
-           log_msg(Test_BI_ADD_FAIL);
-           log_msg(result_msg);
-           return Test_FAIL;
-       }
-
-       memset(str, 0, (test_max_word_size * 8) * 4 + 100); // str 초기화
-       // bigint 뺄셈 테스트
-//       result_msg = test_bi_sub(&op_total_time[1], &nums[0], &nums[1], str);
-       result_msg = test_bi_sub(&op_total_time[1], &a, &b, str);
-       if(result_msg != Test_BI_SUB_SUCCESS){
-           log_msg(Test_BI_SUB_FAIL);
-           log_msg(result_msg);
-           return Test_FAIL;
-       }
-
-       memset(str, 0, (test_max_word_size * 8) * 4 + 100); // str 초기화
-       // BINARY LONG bigint 나눗셈 테스트
-       DIVISION_METHOD = 0;
-//       result_msg = test_bi_div(&op_total_time[2], &nums[0], &nums[1], str, &DIVISION_METHOD);
-       result_msg = test_bi_div(&op_total_time[2], &a, &b, str, &DIVISION_METHOD);
-       if(result_msg != Test_BI_DIV_SUCCESS){
-           log_msg(Test_BI_DIV_FAIL);
-           log_msg(result_msg);
-           return Test_FAIL;
-       }
-
-       memset(str, 0, (test_max_word_size * 8) * 4 + 100); // str 초기화
-       // WORD LONG bigint 나눗셈 테스트
-       DIVISION_METHOD = 1;
-//       result_msg = test_bi_div(&op_total_time[3], &nums[0], &nums[1], str, &DIVISION_METHOD);
-       result_msg = test_bi_div(&op_total_time[3], &a, &b, str, &DIVISION_METHOD);
-       if(result_msg != Test_BI_DIV_SUCCESS){
-           log_msg(Test_BI_DIV_FAIL);
-           log_msg(result_msg);
-           return Test_FAIL;
-       }
-
-       memset(str, 0, (test_max_word_size * 8) * 4 + 100); // str 초기화
-       // bigint 곱셈 테스트
-//       result_msg = test_bi_mul(&op_total_time[4], &nums[0], &nums[1], str);
-       result_msg = test_bi_mul(&op_total_time[4], &a, &b, str);
-       if(result_msg != Test_BI_MUL_SUCCESS){
-           log_msg(Test_BI_MUL_FAIL);
-           log_msg(result_msg);
-           return Test_FAIL;
-       }
-
-       memset(str, 0, (test_max_word_size * 8) * 4 + 100); // str 초기화
-       // bigint 카라츄바 곱셈 테스트
-       karachuba_flag = test_max_word_size / mul_karachuba_ratio;
-//       result_msg = test_bi_mul_karachuba(&op_total_time[5], &nums[0], &nums[1], str, &karachuba_flag);
-       result_msg = test_bi_mul_karachuba(&op_total_time[5], &a, &b, str, &karachuba_flag);
-       if(result_msg != Test_BI_MUL_KARACHUBA_SUCCESS){
-           log_msg(Test_BI_MUL_KARACHUBA_FAIL);
-           log_msg(result_msg);
-           return Test_FAIL;
-       }
-
-       memset(str, 0, (test_max_word_size * 8) * 4 + 100); // str 초기화
-       // bigint 제곱 테스트
-//       result_msg = test_bi_squ(&op_total_time[6], &nums[0], str);
-       result_msg = test_bi_squ(&op_total_time[6], &a, str);
-       if(result_msg != Test_BI_SQU_SUCCESS){
-           log_msg(Test_BI_SQU_FAIL);
-           log_msg(result_msg);
-           return Test_FAIL;
-       }
-
-       squ_karachuba_flag = test_max_word_size / squ_karachuba_ratio;
-       memset(str, 0, (test_max_word_size * 8) * 4 + 100); // str 초기화
-       // bigint 카라츄바 제곱 테스트
-//       result_msg = test_bi_squ_karachuba(&op_total_time[7], &nums[0], str, &squ_karachuba_flag);
-       result_msg = test_bi_squ_karachuba(&op_total_time[7], &a, str, &squ_karachuba_flag);
-       if(result_msg != Test_BI_SQU_KARACHUBA_SUCCESS){
-           log_msg(Test_BI_SQU_KARACHUBA_FAIL);
-           log_msg(result_msg);
-           return Test_FAIL;
-       }
-
-       memset(str, 0, (test_max_word_size * 8) * 4 + 100); // str 초기화
-       // bigint 지수승 테스트
-//       result_msg = test_bi_exp(op_exp_time, &nums[0], &nums[1], &nums[2], str);
-       result_msg = test_bi_exp(op_exp_time, &a, &b, &c, str);
-       if(result_msg != Test_BI_EXP_SUCCESS){
-           return Test_FAIL;
-       }
-*/
-       if(test_word_size == barret_word_size){ // 사전 연산 값이 고정되어 있기에 test_word_size가 기존 사이즈와 같을 때만 수행
-           memset(str, 0, (test_word_size * 8) * 4 + 100); // str 초기화
-           // bigint Barrett Reduction 테스트
-//           result_msg = test_bi_barrett_reduction(&op_total_time[8], &nums[0], &barret_N, &barret_T, str);
-           result_msg = test_bi_barrett_reduction(&op_total_time[8], &a, &barret_N, &barret_T, str);
-           if(result_msg != Test_BI_BARRETT_REDUCTION_SUCCESS){
-               log_msg(Test_BI_BARRETT_REDUCTION_FAIL);
-               log_msg(result_msg);
-               return Test_FAIL;
-           }
-       }
-/*
-       memset(str, 0, (test_max_word_size * 8) * 4 + 100); // str 초기화
-       // bigint gcd 테스트
-//       result_msg = test_bi_gcd(&op_total_time[9], &nums[0], &nums[1], str);
-       result_msg = test_bi_gcd(&op_total_time[9], &a, &b, str);
-       if(result_msg != Test_BI_GCD_SUCCESS){
-           log_msg(Test_BI_GCD_FAIL);
-           log_msg(result_msg);
-           return Test_FAIL;
-       }
-
-       // bigint EEA 테스트
-//       nums[0]->sign = 0;
-//       nums[1]->sign = 0;
-       a->sign = 0;
-       b->sign = 0;
-       memset(str, 0, (test_max_word_size * 8) * 4 + 100); // str 초기화
-       // bigint EEA 테스트
-//       result_msg = test_bi_EEA(&op_total_time[10], &nums[0], &nums[1], str);
-       result_msg = test_bi_EEA(&op_total_time[10], &a, &b, str);
-       if(result_msg != Test_BI_EEA_SUCCESS){
-           log_msg(Test_BI_EEA_FAIL);
-           log_msg(result_msg);
-           return Test_FAIL;
-       }
-*/
-////       nums[0]->sign = 0;
-//       a->sign = 0;
-//       memset(str, 0, (test_max_word_size * 8) * 4 + 100); // str 초기화
-////       result_msg = test_miller_rabin(&op_total_time[11], &nums[0], &miller_rabin_iter, str);
-//       result_msg = test_miller_rabin(&op_total_time[11], &a, &miller_rabin_iter, str);
-//       if(result_msg != Test_MILLER_RABIN_SUCCESS){
-//           log_msg(Test_MILLER_RABIN_FAIL);
-//           log_msg(result_msg);
-//           return Test_FAIL;
-//       }
-   }
-
-//    str = (char *)calloc((test_max_word_size * 8) * 4 + 100, sizeof(char)); // '0x' * 3 + 부호 * 3 + " + " + " = " => 6 + 3 + 3 + 3 = 15
-
-//    fp = fopen("./test_vec/test_bigint.txt", "r");
-//    bigint *nums[3];
-//    int count = 0;
-//    int flag = 1;
-//
-//    while (fgets(str, test_word_size * WORD_BITS, fp) != NULL) {
-//        str[strcspn(str, "\n")] = '\0';
-//        int is_negative = (str[0] == '-');
-//        char *hex_str = is_negative ? str + 3 : str + 2;
-//
-//        bi_set_from_string(&nums[count], hex_str, 16);
-//        if(is_negative) nums[count]->sign = 1;
-//
-//        count++;
-//
-//        // 3개 읽으면 처리 시작
-//        if (count == 3) {
-//            printf("%dth test\n", flag);
-////            if(flag == 50) break;
-//
-//            // 카운터 초기화
-//            count = 0;
-//            flag++;
-//        }
-//    }
-
-//    fclose(fp);
-
-/*
-    printf("\n--------------- WORD INFO ---------------");
-    printf("\nBigint bit size : %d", BITS_SIZE);
-    printf("\nWord Bit Size : %d, Word size : %d", WORD_BITS, test_word_size);
-    printf("\nitersize : %d", test_size);
-    printf("\n============ Testing bi_add ============\n");
-    printf("Time taken add : %f seconds\n", op_total_time[0] / test_size);
-
-//    printf("\n--------------- WORD INFO ---------------");
-//    printf("\nBigint bit size : %d", BITS_SIZE);
-//    printf("\nWord Bit Size : %d, Word size : %d", WORD_BITS, test_word_size);
-//    printf("\nitersize : %d", test_size);
-    printf("\n============ Testing bi_sub ============\n");
-    printf("Time taken sub : %f seconds\n", op_total_time[1] / test_size);
-
-//    printf("\n--------------- WORD INFO ---------------");
-//    printf("\nBigint bit size : %d", BITS_SIZE);
-//    printf("\nWord Bit Size : %d, Word size : %d", WORD_BITS, test_word_size);
-//    printf("\nitersize : %d", test_size);
-    printf("\n============ Testing bi_div (Binary Long Division) ============\n");
-    printf("Time taken div(Binary) : %f seconds\n", op_total_time[2] / test_size);
-
-//    printf("\n--------------- WORD INFO ---------------");
-//    printf("\nBigint bit size : %d", BITS_SIZE);
-//    printf("\nWord Bit Size : %d, Word size : %d", WORD_BITS, test_word_size);
-//    printf("\nitersize : %d", test_size);
-    printf("\n============ Testing bi_div (WORD Long Division) ============\n");
-    printf("Time taken div(Long) : %f seconds\n", op_total_time[3] / test_size);
-
-//    printf("\n--------------- WORD INFO ---------------");
-//    printf("\nBigint bit size : %d", BITS_SIZE);
-//    printf("\nWord Bit Size : %d, Word size : %d", WORD_BITS, test_word_size);
-//    printf("\nitersize : %d", test_size);
-    printf("\n============ Testing bi_mul ============\n");
-    printf("Time taken mul : %f seconds\n", op_total_time[4] / test_size);
-
-//    printf("\n--------------- WORD INFO ---------------");
-//    printf("\nBigint bit size : %d", BITS_SIZE);
-//    printf("\nWord Bit Size : %d, Word size : %d", WORD_BITS, test_word_size);
-//    printf("\nitersize : %d", test_size);
-    printf("\n============ Testing bi_mul_karachuba ============\n");
-    printf("Time taken mul_karachuba : %f seconds\n", op_total_time[5] / test_size);
-
-//    printf("\n--------------- WORD INFO ---------------");
-//    printf("\nBigint bit size : %d", BITS_SIZE);
-//    printf("\nWord Bit Size : %d, Word size : %d", WORD_BITS, test_word_size);
-//    printf("\nitersize : %d", test_size);
-    printf("\n============ Testing bi_squ ============\n");
-    printf("Time taken squ : %f seconds\n", op_total_time[6] / test_size);
-
-//    printf("\n--------------- WORD INFO ---------------");
-//    printf("\nBigint bit size : %d", BITS_SIZE);
-//    printf("\nWord Bit Size : %d, Word size : %d", WORD_BITS, test_word_size);
-//    printf("\nitersize : %d", test_size);
-    printf("\n============ Testing bi_squ_karachuba ============\n");
-    printf("Time taken squ_karachuba : %f seconds\n", op_total_time[7] / test_size);
-
-//    printf("\n--------------- WORD INFO ---------------");
-//    printf("\nBigint bit size : %d", BITS_SIZE);
-//    printf("\nWord Bit Size : %d, Word size : %d", WORD_BITS, test_word_size);
-//    printf("\nitersize : %d", test_size);
-    printf("\n============ Testing bi_exp ============\n");
-    printf("Time taken exp (MS) : %f seconds\n", op_exp_time[0] / test_size);
-    printf("Time taken exp (R TO L) : %f seconds\n", op_exp_time[1] / test_size);
-    printf("Time taken exp (L TO R) : %f seconds\n", op_exp_time[2] / test_size);
-*/
-    printf("\n--------------- WORD INFO ---------------");
-    printf("\nBigint bit size : %d", BITS_SIZE);
-    printf("\nWord Bit Size : %d, Word size : %d", WORD_BITS, test_word_size);
-    printf("\nitersize : %d", test_size);
-    printf("\n============ Testing bi_barrett_reduction ============\n");
-    printf("Time taken barret_reduction : %f seconds\n", op_total_time[8] / test_size);
-
-//
-////    printf("\n--------------- WORD INFO ---------------");
-////    printf("\nBigint bit size : %d", BITS_SIZE);
-////    printf("\nWord Bit Size : %d, Word size : %d", WORD_BITS, test_word_size);
-////    printf("\nitersize : %d", test_size);
-//    printf("\n============ Testing bi_gcd ============\n");
-//    printf("Time taken gcd : %f seconds\n", op_total_time[9] / test_size);
-//
-////    printf("\n--------------- WORD INFO ---------------");
-////    printf("\nBigint bit size : %d", BITS_SIZE);
-////    printf("\nWord Bit Size : %d, Word size : %d", WORD_BITS, test_word_size);
-////    printf("\nitersize : %d", test_size);
-//    printf("\n============ Testing bi_EEA ============\n");
-//    printf("Time taken EEA : %f seconds\n", op_total_time[10] / test_size);
-//    printf("\n");
-
-//    printf("\n============ Testing Miller Rabin ============\n");
-//    printf("Time taken Miller Rabin : %f seconds\n", op_total_time[11] / test_size);
-//    printf("\n");
-//    if(compare_multiplicaiton(16, 120, 16) != COMPARE_MULTIPLICATION_SUCCESS)   return Test_FAIL;   // bigint 곱셈 성능 비교 테스트
-//    if(compare_squaring(16, 120, 16) != COMPARE_SQUARING_SUCCESS)   return Test_FAIL;   // bigint 곱셈 성능 비교 테스트
-//    if(compare_division(16, 120, 16) != COMPARE_DIVISION_SUCCESS)   return Test_FAIL;   // bigint 곱셈 성능 비교 테스트
-
-    // 카라츄바 세팅 해제
-    if(clear_karachuba_pool() != CLEAR_KARACHUBA_POOL_SUCCESS){
-        log_msg(CLEAR_KARACHUBA_POOL_FAIL);
-        return Test_FAIL;
     }
-
-    printf("\n");
-    log_msg(Test_SUCCESS);
-
-    // Sage test
-    fp = popen("python3 ../../sage_test/test.py >/dev/null 2>&1", "r");
-    if (fp == NULL){
-        printf("Failed to run command\n");
-        return 1;
-    }
-
-    // Close the command stream
-    int status = pclose(fp);
-    if (status == -1){
-        perror("pclose failed");
-        return 1;
-    }
-    else
-        printf("python script exited with status %d\n", status);
-
-    return 0;
-TEST_EXIT:
-    if(bi_delete(&a) != BI_FREE_SUCCESS)    return Test_FAIL;
-    if(bi_delete(&b) != BI_FREE_SUCCESS)    return Test_FAIL;
-    if(bi_delete(&c) != BI_FREE_SUCCESS)    return Test_FAIL;
-
-    log_msg(result_msg);
-    return Test_FAIL;
+    result_msg = new_bigint_test_vector_SUCCESS;
+NEW_BIGINT_EXIT:
+    if (str != NULL)    free(str);
+    if (a != NULL)    bi_delete(&a);
+    return result_msg;
 }
 
 msg test_bi_new_delete(){
@@ -989,698 +593,4 @@ EXIT_RANDOM:
     log_msg(result_msg);
     if (Test_file_write(Test_main_file, seperator, APPEND) != FILE_WRITE_SUCCESS)   return FILE_WRITE_FAIL;
     return Test_SUCCESS;
-}
-
-msg test_bi_add(OUT double* total_time_add, IN bigint** a, IN bigint** b, IN char* str){
-    bigint *c = NULL;
-    msg result_msg = Test_BI_ADD_SUCCESS;
-    ParamType param_types[2] = {TYPE_BIGINT_PTR, TYPE_BIGINT_PTR};
-
-    if (bigint_to_hex(str, a) == -1)   goto ADD_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_add, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto ADD_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_add, " + ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto ADD_EXIT;
-
-    if (bigint_to_hex(str, b) == -1)   goto ADD_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_add, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto ADD_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_add, " = ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto ADD_EXIT;
-
-    *total_time_add += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_add, &c, &result_msg, param_types, a, b);
-    if (result_msg != BI_ADD_SUCCESS)   goto ADD_EXIT;
-
-    if (bigint_to_hex(str, &c) == -1)   goto ADD_EXIT;
-    result_msg = Test_file_write(Test_file_add, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto ADD_EXIT;
-
-    result_msg = Test_BI_ADD_SUCCESS;
-
-ADD_EXIT:
-    if (bi_delete(&c) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
-}
-
-msg test_bi_sub(OUT double* total_time_sub, IN bigint** a, IN bigint** b, IN char* str){
-    bigint *c = NULL;
-    msg result_msg = Test_BI_SUB_SUCCESS;
-    ParamType param_types[2] = {TYPE_BIGINT_PTR, TYPE_BIGINT_PTR};
-
-    if (bigint_to_hex(str, a) == -1)   goto SUB_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_sub, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto SUB_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_sub, " - ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto SUB_EXIT;
-
-    if (bigint_to_hex(str, b) == -1)   goto SUB_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_sub, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto SUB_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_sub, " = ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto SUB_EXIT;
-
-    *total_time_sub += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_sub, &c, &result_msg, param_types, a, b);
-    if (result_msg != BI_SUB_SUCCESS)   goto SUB_EXIT;
-
-    if (bigint_to_hex(str, &c) == -1)   goto SUB_EXIT;
-    result_msg = Test_file_write(Test_file_sub, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto SUB_EXIT;
-
-    result_msg = Test_BI_SUB_SUCCESS;
-
-SUB_EXIT:
-    if (bi_delete(&c) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
-}
-
-msg test_bi_mul(OUT double* total_time_mul, IN bigint** a, IN bigint** b, IN char* str){
-    bigint *c = NULL;
-    msg result_msg = Test_BI_MUL_SUCCESS;
-    ParamType param_types[2] = {TYPE_BIGINT_PTR, TYPE_BIGINT_PTR};
-
-    if (bigint_to_hex(str, a) == -1)   goto MUL_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_mul, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto MUL_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_mul, " * ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto MUL_EXIT;
-
-    if (bigint_to_hex(str, b) == -1)   goto MUL_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_mul, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto MUL_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_mul, " = ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto MUL_EXIT;
-
-    *total_time_mul += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_mul, &c, &result_msg, param_types, a, b);
-    if (result_msg != BI_MUL_SUCCESS)   goto MUL_EXIT;
-
-    if (bigint_to_hex(str, &c) == -1)   goto MUL_EXIT;
-    result_msg = Test_file_write(Test_file_mul, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto MUL_EXIT;
-
-    result_msg = Test_BI_MUL_SUCCESS;
-
-MUL_EXIT:
-    if (bi_delete(&c) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
-}
-
-
-msg test_bi_mul_karachuba(OUT double* total_time_mul_karachuba, IN bigint** a, IN bigint** b, IN char* str, IN int* karachuba_flag){
-    bigint *c = NULL;
-    msg result_msg = Test_BI_MUL_KARACHUBA_SUCCESS;
-    ParamType param_types[3] = {TYPE_BIGINT_PTR, TYPE_BIGINT_PTR, TYPE_INT_PTR};
-
-    if (bigint_to_hex(str, a) == -1)   goto MUL_KARACHUBA_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_mul_karachuba, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto MUL_KARACHUBA_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_mul_karachuba, " * ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto MUL_KARACHUBA_EXIT;
-
-    if (bigint_to_hex(str, b) == -1)   goto MUL_KARACHUBA_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_mul_karachuba, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto MUL_KARACHUBA_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_mul_karachuba, " = ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto MUL_KARACHUBA_EXIT;
-
-    *total_time_mul_karachuba += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_mul_karachuba, &c, &result_msg, param_types, a, b, karachuba_flag);
-    if (result_msg != BI_MUL_SUCCESS)   goto MUL_KARACHUBA_EXIT;
-
-    if (bigint_to_hex(str, &c) == -1)   goto MUL_KARACHUBA_EXIT;
-    result_msg = Test_file_write(Test_file_mul_karachuba, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto MUL_KARACHUBA_EXIT;
-
-    result_msg = Test_BI_MUL_KARACHUBA_SUCCESS;
-
-MUL_KARACHUBA_EXIT:
-    if (bi_delete(&c) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
-}
-
-msg compare_multiplicaiton(int start_size, int end_size, int step_size){
-    printf("\n=== Comparing Multiplication Methods ===\n");
-    printf("Size\titeration\tClassic(s)\tKaratsuba(s)\tRatio\tCrossover\n");
-    printf("------------------------------------------------------------------------------\n");
-
-    bigint *a = NULL;
-    bigint *b = NULL;
-    bigint *c = NULL;
-    msg result_msg = COMPARE_MULTIPLICATION_SUCCESS;
-    int crossover_found = 0;
-    int crossover_point = 0;
-    int karachuba_flag = 0;
-    ParamType param_types[2] = {TYPE_BIGINT_PTR, TYPE_BIGINT_PTR};
-    ParamType param_types_karachuba[3] = {TYPE_BIGINT_PTR, TYPE_BIGINT_PTR, TYPE_INT_PTR};
-
-    for (int word_size = start_size; word_size <= end_size; word_size += step_size){
-        double total_time_classic = 0;
-        double total_time_karatsuba = 0;
-
-        for (int i = 0; i < test_size; i++){
-            result_msg = bi_get_random(&a, word_size);
-            if (result_msg != BI_GET_RANDOM_SUCCESS)
-                goto COMAPARE_MUL_EXIT;
-            else if(a->word_len != word_size){
-                result_msg = BI_GET_RANDOM_LENGTH_NOT_MATCH;
-                goto COMAPARE_MUL_EXIT;
-            }
-
-            result_msg = bi_get_random(&b, word_size);
-            if (result_msg != BI_GET_RANDOM_SUCCESS)
-                goto COMAPARE_MUL_EXIT;
-            else if(b->word_len != word_size){
-                result_msg = BI_GET_RANDOM_LENGTH_NOT_MATCH;
-                goto COMAPARE_MUL_EXIT;
-            }
-
-            total_time_classic += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_mul, &c, &result_msg, param_types, &a, &b);
-            if (result_msg != BI_MUL_SUCCESS)
-                goto COMAPARE_MUL_EXIT;
-
-            karachuba_flag = word_size / mul_karachuba_ratio;
-            total_time_karatsuba += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_mul_karachuba, &c, &result_msg, param_types_karachuba, &a, &b, &karachuba_flag);
-            if (result_msg != BI_MUL_SUCCESS)
-                goto COMAPARE_MUL_EXIT;
-        }
-        double avg_time_classic = total_time_classic / test_size;
-        double avg_time_karatsuba = total_time_karatsuba / test_size;
-        double ratio = avg_time_classic / avg_time_karatsuba;
-
-        // Crossover point 찾기 (Karatsuba가 더 빨라지는 지점)
-        if (!crossover_found && ratio > 1.0){
-            crossover_found = 1;
-            crossover_point = word_size;
-        }
-
-        printf("%d\t%d\t\t%.6f\t%.6f\t%.2f\t%s\n",
-               word_size,
-               test_size,
-               avg_time_classic,
-               avg_time_karatsuba,
-               ratio,
-               (word_size == crossover_point) ? "<=== Crossover" : "");
-    }
-
-    if (crossover_found){
-        printf("\nKaratsuba becomes faster at word size: %d\n\n", crossover_point);
-    }
-    result_msg = COMPARE_MULTIPLICATION_SUCCESS;
-
-COMAPARE_MUL_EXIT:
-    if (bi_delete(&a) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    if (bi_delete(&b) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    if (bi_delete(&c) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
-}
-
-msg compare_squaring(int start_size, int end_size, int step_size){
-    printf("\n=== Comparing Squaring Methods ===\n");
-    printf("Size\titeration\tClassic(s)\tKaratsuba(s)\tRatio\tCrossover\n");
-    printf("------------------------------------------------------------------------------\n");
-
-    bigint *a = NULL;
-    bigint *c = NULL;
-    msg result_msg = COMPARE_SQUARING_SUCCESS;
-    int crossover_found = 0;
-    int crossover_point = 0;
-    int squ_karachuba_flag = 0;
-    ParamType param_types[1] = {TYPE_BIGINT_PTR};
-    ParamType param_types_karachuba[2] = {TYPE_BIGINT_PTR, TYPE_INT_PTR};
-
-    for (int word_size = start_size; word_size <= end_size; word_size += step_size){
-        double total_time_classic = 0;
-        double total_time_karatsuba = 0;
-
-        for (int i = 0; i < test_size; i++){
-            result_msg = bi_get_random(&a, word_size);
-            if (result_msg != BI_GET_RANDOM_SUCCESS)
-                goto COMAPARE_SQU_EXIT;
-            else if(a->word_len != word_size){
-                result_msg = BI_GET_RANDOM_LENGTH_NOT_MATCH;
-                goto COMAPARE_SQU_EXIT;
-            }
-
-            total_time_classic += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_squ, &c, &result_msg, param_types, &a);
-            if (result_msg != BI_SQU_SUCCESS)
-                goto COMAPARE_SQU_EXIT;
-
-            squ_karachuba_flag = word_size / squ_karachuba_ratio;
-            total_time_karatsuba += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_squ_karachuba, &c, &result_msg, param_types_karachuba, &a, &squ_karachuba_flag);
-            if (result_msg != BI_SQU_SUCCESS)
-                goto COMAPARE_SQU_EXIT;
-        }
-        double avg_time_classic = total_time_classic / test_size;
-        double avg_time_karatsuba = total_time_karatsuba / test_size;
-        double ratio = avg_time_classic / avg_time_karatsuba;
-
-        // Crossover point 찾기 (Karatsuba가 더 빨라지는 지점)
-        if (!crossover_found && ratio > 1.0){
-            crossover_found = 1;
-            crossover_point = word_size;
-        }
-
-        printf("%d\t%d\t\t%.6f\t%.6f\t%.2f\t%s\n",
-               word_size,
-               test_size,
-               avg_time_classic,
-               avg_time_karatsuba,
-               ratio,
-               (word_size == crossover_point) ? "<=== Crossover" : "");
-    }
-
-    if (crossover_found){
-        printf("\nKaratsuba becomes faster at word size: %d\n\n", crossover_point);
-    }
-    result_msg = COMPARE_SQUARING_SUCCESS;
-
-COMAPARE_SQU_EXIT:
-    if (bi_delete(&a) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    if (bi_delete(&c) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
-}
-
-msg compare_division(int start_size, int end_size, int step_size){
-    printf("\n=== Comparing Division Methods ===\n");
-    printf("Size\titeration\tbinary Long(s)\tWord Long(s)\tRatio\tCrossover\n");
-    printf("------------------------------------------------------------------------------\n");
-
-    bigint *a = NULL;
-    bigint *b = NULL;
-    bigint *c = NULL;
-    bigint *d = NULL;
-    msg result_msg = COMPARE_DIVISION_SUCCESS;
-    int crossover_found = 0;
-    int crossover_point = 0;
-    int div_option = 0;
-    ParamType param_types[4] = {TYPE_BIGINT_PTR, TYPE_BIGINT_PTR, TYPE_BIGINT_PTR, TYPE_INT_PTR};
-
-    for (int word_size = start_size; word_size <= end_size; word_size += step_size){
-        double total_time_binary = 0;
-        double total_time_word = 0;
-
-        for (int i = 0; i < test_size; i++){
-            result_msg = bi_get_random(&a, word_size);
-            if (result_msg != BI_GET_RANDOM_SUCCESS)
-                goto COMAPARE_DIV_EXIT;
-            else if(a->word_len != word_size){
-                result_msg = BI_GET_RANDOM_LENGTH_NOT_MATCH;
-                goto COMAPARE_DIV_EXIT;
-            }
-
-            result_msg = bi_get_random(&b, word_size);
-            if (result_msg != BI_GET_RANDOM_SUCCESS)
-                goto COMAPARE_DIV_EXIT;
-            else if(b->word_len != word_size){
-                result_msg = BI_GET_RANDOM_LENGTH_NOT_MATCH;
-                goto COMAPARE_DIV_EXIT;
-            }
-            div_option = 0;
-            total_time_binary += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_div, &d, &result_msg, param_types, &c, &a, &b, &div_option);
-            if (result_msg != BI_DIV_SUCCESS)
-                goto COMAPARE_DIV_EXIT;
-
-            div_option = 1;
-            total_time_word += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_div, &d, &result_msg, param_types, &c, &a, &b, &div_option);
-            if (result_msg != BI_DIV_SUCCESS)
-                goto COMAPARE_DIV_EXIT;
-        }
-        double avg_time_binary = total_time_binary / test_size;
-        double avg_time_word = total_time_word / test_size;
-        double ratio = avg_time_binary / avg_time_word;
-
-        // Crossover point 찾기 (Karatsuba가 더 빨라지는 지점)
-        if (!crossover_found && ratio > 1.0){
-            crossover_found = 1;
-            crossover_point = word_size;
-        }
-
-        printf("%d\t%d\t\t%.6f\t%.6f\t%.2f\t%s\n",
-               word_size,
-               test_size,
-               avg_time_binary,
-               avg_time_word,
-               ratio,
-               (word_size == crossover_point) ? "<=== Crossover" : "");
-    }
-
-    if (crossover_found){
-        printf("\nKaratsuba becomes faster at word size: %d\n\n", crossover_point);
-    }
-    result_msg = COMPARE_DIVISION_SUCCESS;
-
-COMAPARE_DIV_EXIT:
-    if (bi_delete(&a) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    if (bi_delete(&b) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    if (bi_delete(&c) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    if (bi_delete(&d) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    log_msg(result_msg);
-    return result_msg;
-}
-
-
-msg test_bi_div(OUT double* total_time_div, IN bigint** a, IN bigint** b, IN char* str, IN int* option){
-    bigint *q = NULL;
-    bigint *r = NULL;
-    msg result_msg = Test_BI_DIV_SUCCESS;
-    ParamType param_types[4] = {TYPE_BIGINT_PTR, TYPE_BIGINT_PTR, TYPE_BIGINT_PTR, TYPE_INT_PTR};
-
-    if (bigint_to_hex(str, a) == -1)   goto DIV_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_div, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto DIV_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_div, " / ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto DIV_EXIT;
-
-    if (bigint_to_hex(str, b) == -1)   goto DIV_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_div, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto DIV_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_div, " = ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto DIV_EXIT;
-
-    *total_time_div += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_div, &q, &result_msg, param_types, &r, a, b, option);
-    if(result_msg == BI_DIV_BY_ZERO){
-        result_msg = Test_file_write(Test_file_div, "DIVISION BY ZERO", APPEND);
-        goto DIV_EXIT;
-    }
-
-    if (result_msg != BI_DIV_SUCCESS)   goto DIV_EXIT;
-
-    if (bigint_to_hex(str, &q) == -1)   goto DIV_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_div, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto DIV_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_div, ", ", APPEND); // 구분자
-    if (result_msg != FILE_WRITE_SUCCESS)   goto DIV_EXIT;
-
-    if (bigint_to_hex(str, &r) == -1)   goto DIV_EXIT;
-    result_msg = Test_file_write(Test_file_div, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto DIV_EXIT;
-
-    result_msg = Test_BI_DIV_SUCCESS;
-
-DIV_EXIT:
-    if (bi_delete(&q) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    if (bi_delete(&r) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
-}
-
-msg test_bi_squ(OUT double* total_time_squ, IN bigint** a, IN char* str){
-    bigint *b = NULL;
-    msg result_msg = Test_BI_SQU_SUCCESS;
-    ParamType param_types[1] = {TYPE_BIGINT_PTR};
-
-    if (bigint_to_hex(str, a) == -1)   goto SQU_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_squ, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)goto SQU_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_squ, " * ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto SQU_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_squ, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto SQU_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_squ, " = ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto SQU_EXIT;
-
-    *total_time_squ += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_squ, &b, &result_msg, param_types, a);
-    if (result_msg != BI_SQU_SUCCESS)   goto SQU_EXIT;
-
-    if (bigint_to_hex(str, &b) == -1)   goto SQU_EXIT;
-    result_msg = Test_file_write(Test_file_squ, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto SQU_EXIT;
-
-    result_msg = Test_BI_SQU_SUCCESS;
-
-SQU_EXIT:
-    if (bi_delete(&b) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
-}
-
-msg test_bi_squ_karachuba(OUT double* total_time_squ_karachuba, IN bigint** a, IN char* str, IN int* squ_karachuba_flag){
-    bigint *b = NULL;
-    msg result_msg = Test_BI_SQU_KARACHUBA_SUCCESS;
-    ParamType param_types[2] = {TYPE_BIGINT_PTR, TYPE_INT_PTR};
-
-    if (bigint_to_hex(str, a) == -1)   goto SQU_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_squ_karachuba, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)goto SQU_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_squ_karachuba, " * ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto SQU_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_squ_karachuba, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto SQU_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_squ_karachuba, " = ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto SQU_EXIT;
-
-    *total_time_squ_karachuba += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_squ_karachuba, &b, &result_msg, param_types, a, squ_karachuba_flag);
-    if (result_msg != BI_SQU_SUCCESS)   goto SQU_EXIT;
-
-    if (bigint_to_hex(str, &b) == -1)   goto SQU_EXIT;
-    result_msg = Test_file_write(Test_file_squ_karachuba, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto SQU_EXIT;
-
-    result_msg = Test_BI_SQU_KARACHUBA_SUCCESS;
-
-SQU_EXIT:
-    if (bi_delete(&b) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
-}
-
-msg test_bi_exp(OUT double total_time_exp[3], IN bigint** a, IN bigint** b, IN bigint** c, IN char* str){
-    bigint *d = NULL; // 결과 bigint
-    msg result_msg = Test_BI_EXP_FAIL;
-    ParamType param_types[3] = {TYPE_BIGINT_PTR, TYPE_BIGINT_PTR, TYPE_BIGINT_PTR};
-
-    (*a)->sign = 0; // 여기도 일단 양수로만 하자
-
-    if (bigint_to_hex(str, a) == -1)   goto EXP_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_exp, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)goto EXP_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_exp, " ^ ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EXP_EXIT;
-
-    (*b)->sign = 0; // 양수를 기반으로 수행
-    (*c)->sign = 0; // 양수를 기반으로 수행
-
-    if (bigint_to_hex(str, b) == -1)   goto EXP_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_exp, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EXP_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_exp, " mod ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EXP_EXIT;
-
-    if(bigint_to_hex(str, c) == -1) goto EXP_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_exp, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EXP_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_exp, " = ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EXP_EXIT;
-
-    // Multipliation Squaring 구현
-    total_time_exp[0] += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_exp_MS, &d, &result_msg, param_types, a, b, c);
-    if (result_msg != BI_EXP_MS_SUCCESS)   goto EXP_EXIT;
-
-    if (bigint_to_hex(str, &d) == -1)   goto EXP_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_exp, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EXP_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_exp, " , ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EXP_EXIT;
-
-    // Right to Left 구현
-    total_time_exp[1] += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_exp_R_TO_L, &d, &result_msg, param_types, a, b, c);
-    if (result_msg != BI_EXP_R_TO_L_SUCCESS)   goto EXP_EXIT;
-
-    if (bigint_to_hex(str, &d) == -1)   goto EXP_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_exp, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EXP_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_exp, " , ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EXP_EXIT;
-
-    // Left to Right 구현
-    total_time_exp[2] += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_exp_L_TO_R, a, &result_msg, param_types, a, b, c);
-    if (result_msg != BI_EXP_L_TO_R_SUCCESS)   goto EXP_EXIT;
-
-    if (bigint_to_hex(str, a) == -1)   goto EXP_EXIT;
-    result_msg = Test_file_write(Test_file_exp, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EXP_EXIT;
-
-    result_msg = Test_BI_EXP_SUCCESS;
-
-EXP_EXIT:
-    if (bi_delete(&d) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
-}
-
-msg test_bi_barrett_reduction(OUT double* total_time_barret_reduction, IN bigint** a, IN bigint** barret_N, IN bigint** barret_T, IN char* str){
-    bigint *b = NULL;
-    msg result_msg = Test_BI_BARRETT_REDUCTION_FAIL;
-
-    ParamType param_types[3] = {TYPE_BIGINT_PTR, TYPE_BIGINT_PTR, TYPE_BIGINT_PTR};
-
-    if (bigint_to_hex(str, a) == -1)   goto BARRET_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_barrett_reduction, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)goto BARRET_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_barrett_reduction, " mod ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto BARRET_EXIT;
-
-    if (bigint_to_hex(str, barret_N) == -1)   goto BARRET_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_barrett_reduction, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)goto BARRET_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_barrett_reduction, " = ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto BARRET_EXIT;
-
-    *total_time_barret_reduction += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())barret_reduction, &b, &result_msg, param_types, a, barret_N, barret_T);
-    if (result_msg != BI_BARRET_REDUCTION_SUCCESS)   goto BARRET_EXIT;
-
-    if (bigint_to_hex(str, &b) == -1)   goto BARRET_EXIT;
-    result_msg = Test_file_write(Test_file_barrett_reduction, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto BARRET_EXIT;
-
-    result_msg = Test_BI_BARRETT_REDUCTION_SUCCESS;
-
-BARRET_EXIT:
-    if (bi_delete(&b) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
-}
-
-
-msg test_bi_gcd(OUT double* total_time_div, IN bigint** a, IN bigint** b, IN char* str){
-    bigint *d = NULL;
-    msg result_msg = Test_BI_GCD_FAIL;
-    ParamType param_types[2] = {TYPE_BIGINT_PTR,TYPE_BIGINT_PTR};
-
-    result_msg = Test_file_write_non_enter(Test_file_gcd, "gcd ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto GCD_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_gcd, "( ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto GCD_EXIT;
-    
-    if (bigint_to_hex(str, a) == -1)   goto GCD_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_gcd, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto GCD_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_gcd, " , ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto GCD_EXIT;
-
-     if (bigint_to_hex(str, b) == -1)   goto GCD_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_gcd, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto GCD_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_gcd, " ) ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto GCD_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_gcd, "= ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto GCD_EXIT;
-   
-    *total_time_div += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_gcd, &d, &result_msg, param_types, a, b);
-    if (result_msg != BI_GCD_SUCCESS)   goto GCD_EXIT;
-
-    if (bigint_to_hex(str, &d) == -1)   goto GCD_EXIT;
-    result_msg = Test_file_write(Test_file_gcd, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto GCD_EXIT;
-
-    result_msg = Test_BI_GCD_SUCCESS;
-
-GCD_EXIT:
-    if (bi_delete(&d) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
-}
-
-msg test_bi_EEA(OUT double* total_time_eea, IN bigint** a, IN bigint** b, IN char* str){
-    bigint *gcd = NULL;
-    bigint *x = NULL;
-    bigint *y = NULL;
-    msg result_msg = Test_BI_EEA_FAIL;
-    ParamType param_types[4] = {TYPE_BIGINT_PTR, TYPE_BIGINT_PTR, TYPE_BIGINT_PTR, TYPE_BIGINT_PTR};
-
-    result_msg = Test_file_write_non_enter(Test_file_EEA, "gcd ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EEA_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_EEA, "( ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EEA_EXIT;
-
-    if (bigint_to_hex(str, a) == -1)   goto EEA_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_EEA, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EEA_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_EEA, " , ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EEA_EXIT;
-
-    if (bigint_to_hex(str, b) == -1)   goto EEA_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_EEA, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EEA_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_EEA, " ) ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EEA_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_EEA, "= ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EEA_EXIT;
-
-    *total_time_eea += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())bi_EEA, &gcd, &result_msg, param_types, &x, &y, a, b);
-    if (result_msg != BI_EEA_SUCCESS)   goto EEA_EXIT;
-
-    if (bigint_to_hex(str, &gcd) == -1)   goto EEA_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_EEA, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EEA_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_EEA, " , ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EEA_EXIT;
-
-    if (bigint_to_hex(str, &x) == -1)   goto EEA_EXIT;
-    result_msg = Test_file_write_non_enter(Test_file_EEA, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EEA_EXIT;
-
-    result_msg = Test_file_write_non_enter(Test_file_EEA, " , ", APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EEA_EXIT;
-
-    if (bigint_to_hex(str, &y) == -1)   goto EEA_EXIT;
-    result_msg = Test_file_write(Test_file_EEA, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EEA_EXIT;
-
-    result_msg = Test_BI_EEA_SUCCESS;
-
-EEA_EXIT:
-    if (bi_delete(&gcd) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    if (bi_delete(&x) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    if (bi_delete(&y) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
-}
-
-msg test_miller_rabin(double* total_time_MR , IN bigint** a, IN int* iteration, char* str){
-    msg result_msg = Test_MILLER_RABIN_FAIL;
-    ParamType param_types[2] = {TYPE_BIGINT_PTR, TYPE_INT_PTR};
-    bigint* temp = NULL;
-
-    *total_time_MR += CHECK_FUNCTION_RUN_ONE_TIME((msg (*)())miller_rabin_primality, &temp, &result_msg, param_types, a, iteration);
-    if (result_msg != MR_SUCCESS)   goto EXIT_MR;
-
-    if(temp->a[0])  result_msg = Test_file_write_non_enter(Test_file_MR, "Probably_Prime ", APPEND);
-    else    result_msg = Test_file_write_non_enter(Test_file_MR, "Composite ", APPEND);
-
-    if (bigint_to_hex(str, a) == -1)   goto EXIT_MR;
-    result_msg = Test_file_write(Test_file_MR, str, APPEND);
-    if (result_msg != FILE_WRITE_SUCCESS)   goto EXIT_MR;
-
-    result_msg = bi_delete(&temp);
-    if (result_msg != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return Test_MILLER_RABIN_SUCCESS;
-
-EXIT_MR:
-    if (bi_delete(&temp) != BI_FREE_SUCCESS)   return BI_FREE_FAIL;
-    return result_msg;
 }
