+++ ref/test/util_test.h
@@ -14,12 +14,7 @@
 #include "../data_type.h"
 #include "../file_io.h"
 #include "../crypto.h"
-
-#define BITS_SIZE 1024                              // 테스트 비트 크기
-#define test_word_size BITS_SIZE / WORD_BITS               // 0인 경우 테스트 시 사용되는 word 크기는 랜덤
-#define test_word_size_limit BITS_SIZE / WORD_BITS          // 랜덤 word 크기 사용 시 제한 범위
-#define test_size  1000                                   // 테스트 횟수
-#define barret_word_size test_word_size                // barret N 크기
+#include "random_test.h"
 
 /**
  * @brief Test bigint structure allocation and deallocation
@@ -50,53 +45,6 @@ msg test_bi_set_from_base(IN int* base);
 msg test_bi_random();
 
 /**
- * @brief Test bigint structure Add operation
- *
- * @return message SUCCESS or FAIL
- */
-msg test_bi_add(OUT double* total_time_add, IN bigint** a, IN bigint** b, IN char* str);
-
-/**
- * @brief Test bigint structure Sub operation
- *
- * @return message SUCCESS or FAIL
- */
-msg test_bi_sub(OUT double* total_time_sub, IN bigint** a, IN bigint** b, IN char* str);
-
-/**
- * @brief Test bigint structure Mul operation
- *
- * @return message SUCCESS or FAIL
- */
-msg test_bi_mul(OUT double* total_time_mul, IN bigint** a, IN bigint** b, IN char* str);
-
-/**
- * @brief Test bigint structure Mul operation using Karachuba algorithm
- *
- * @return message SUCCESS or FAIL
- */
-msg test_bi_mul_karachuba(OUT double* total_time_mul_karachuba, IN bigint** a, IN bigint** b, IN char* str, IN int* karachuba_flag);
-
-/**
- * @brief Test classic and Karachuba algorithm same value
- *
- * @param start_size start size of bigint
- * @param end_size end size of bigint
- * @param step_size step size of bigint
- * @return message SUCCESS or FAIL
- */
-msg compare_multiplicaiton(int start_size, int end_size, int step_size);
-msg compare_squaring(int start_size, int end_size, int step_size);
-msg compare_division(int start_size, int end_size, int step_size);
-
-/**
- * @brief Test bigint structure Division operation
- *
- * @return message SUCCESS or FAIL
- */
-msg test_bi_div(OUT double* total_time_div, IN bigint** a, IN bigint** b, IN char* str, IN int* option);
-
-/**
  * @brief Test bigint structure Shift operation
  *
  * @return message SUCCESS or FAIL
@@ -118,52 +66,9 @@ msg test_bi_get_lower();
 msg test_bi_cat();
 
 /**
- * @brief Test bigint structure Squaring operation
+ * @brief Create new bigint test vector
  *
  * @return message SUCCESS or FAIL
  */
-msg test_bi_squ(OUT double* total_time_squ, IN bigint** a, IN char* str);
-
-/**
- * @brief Test bigint structure Squaring operation by using Karachuba algorithm
- *
- * @return message SUCCESS or FAIL
- */
-msg test_bi_squ_karachuba(OUT double* total_time_squ_karachuba, IN bigint** a, IN char* str, IN int* squ_karachuba_flag);
-
-/**
- * @brief Test bigint structure Exponentional operation
- *
- * @return message SUCCESS or FAIL
- */
-msg test_bi_exp(OUT double total_time_exp[3], IN bigint** a, IN bigint** b, IN bigint** c, IN char* str);
-
-/**
- * @brief Test bigint structure Barret Reduction operation
- *
- * @return message SUCCESS or FAIL
- */
-msg test_bi_barrett_reduction(OUT double* total_time_barret_reduction, IN bigint** a, IN bigint** barret_N, IN bigint** barret_T, IN char* str);
-
-/**
- * @brief Test bigint structure GCD operation
- *
- * @return message SUCCESS or FAIL
- */
-msg test_bi_gcd(OUT double* total_time_div, IN bigint** a, IN bigint** b, IN char* str);
-
-/**
- * @brief Test bigint structure Extended Eucliean Algorithm operation
- *
- * @return message SUCCESS or FAIL
- */
-msg test_bi_EEA(OUT double* total_time_div, IN bigint** a, IN bigint** b, IN char* str);
-
-
-/**
- * @brief Test bigint structure Miller Rabin operation
- *
- * @return message SUCCESS or FAIL
- */
-msg test_miller_rabin(double* total_time_MR , IN bigint** a, IN int* iteration, char* str);
+msg new_bigint_test_vector();
 #endif // UTIL_TEST_H
