+++ sage_test/test.py
@@ -16,8 +16,6 @@ def addition(add_str):
 
 def test_addition(f, p):
     k = open('./result/result_addition.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 덧셈 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -31,7 +29,10 @@ def test_addition(f, p):
             k.write(addif)
             k.write(str(hex(tmp)))
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 덧셈 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def subtraction(sub_str):
     sub_str = sub_str.split(' ')
@@ -43,8 +44,6 @@ def subtraction(sub_str):
 
 def test_subtraction(f, p):
     k = open('./result/result_subtraction.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 뺄셈 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -58,7 +57,10 @@ def test_subtraction(f, p):
             k.write(subif)
             k.write(str(hex(tmp)))
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 뺄셈 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def multiplicaiton(sub_str):
     sub_str = sub_str.split(' ')
@@ -70,8 +72,6 @@ def multiplicaiton(sub_str):
 
 def test_multiplication(f, p):
     k = open('./result/result_multiplication.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 곱셈 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -85,7 +85,10 @@ def test_multiplication(f, p):
             k.write(subif)
             k.write(str(hex(tmp)))
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 곱셈 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def string_to_hex(p, string, result, base):
     hex_string = int(string, base)
@@ -97,7 +100,6 @@ def string_to_hex(p, string, result, base):
 def test_bi_set_from_loop(f, p, k, base):
     count = 0
     False_count = 0
-    p.write(f'[{base}진수]\n')
     k.write(f'[{base}진수]\n')
     while True:
         string = f.readline()
@@ -112,7 +114,9 @@ def test_bi_set_from_loop(f, p, k, base):
             k.write(f"올바른 값 : {tmp}\n")
             k.write(f"제시 값   : {result}")
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write(f'[{base}진수]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def test_bi_set_from(f, p):
     k = open('./result/result_set_from.txt', 'w')
@@ -131,8 +135,6 @@ def test_bi_set_from(f, p):
 
 def test_karachuba_multiplication(f, p):
     k = open('./result/result_multiplication_karachuba.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[카라츄바 빅넘 곱셈 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -146,7 +148,10 @@ def test_karachuba_multiplication(f, p):
             k.write(subif)
             k.write(str(hex(tmp)))
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[카라츄바 빅넘 곱셈 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def division(sub_str):
     sub_str = sub_str.split(' ')
@@ -167,8 +172,6 @@ def division(sub_str):
 
 def test_division(f, p):
     k = open('./result/result_division.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 나눗셈 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -184,12 +187,13 @@ def test_division(f, p):
             k.write(', ')
             k.write(str(hex(r)))
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 나눗셈 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def test_shift(f, p):
     k = open('./result/result_shift.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 시프트 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -208,12 +212,13 @@ def test_shift(f, p):
             k.write(shiftif)
             k.write(str(hex(tmp)))
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 시프트 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def test_get_lower(f, p):
     k = open('./result/result_get_lower.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 and 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -240,12 +245,13 @@ def test_get_lower(f, p):
             k.write(shiftif)
             k.write(str(hex(tmp)))
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 and 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def test_cat(f, p):
     k = open('./result/result_cat.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 cat 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -261,12 +267,13 @@ def test_cat(f, p):
             k.write(shiftif)
             k.write(str(tmp))
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 cat 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def test_squ(f, p):
     k = open('./result/result_seq.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 제곱 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -282,13 +289,14 @@ def test_squ(f, p):
             k.write(shiftif)
             k.write(str(hex(tmp)))
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 제곱 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 
 def test_karachuba_squ(f, p):
     k = open('./result/result_seq_karachuba.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 카라츄바 제곱 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -304,12 +312,13 @@ def test_karachuba_squ(f, p):
             k.write(shiftif)
             k.write(str(hex(tmp)))
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 카라츄바 제곱 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def test_exp(f, p):
     k = open('./result/result_exp.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 모듈러 지수승 연산]\n')
     count = 0
     False_count_MS = 0
     False_count_R_TO_L = 0
@@ -334,14 +343,15 @@ def test_exp(f, p):
             k.write(shiftif)
             k.write(str(hex(tmp)))
             k.write('\n')
-    p.write(f"[Multiplication and Squaring]\n실행 횟수 : {count} / 성공 횟수 : {count - False_count_MS} / 실패 횟수 : {False_count_MS}\n")
-    p.write(f"[Right to Left]\n실행 횟수 : {count} / 성공 횟수 : {count - False_count_R_TO_L} / 실패 횟수 : {False_count_R_TO_L}\n")
-    p.write(f"[Left to Right]\n실행 횟수 : {count} / 성공 횟수 : {count - False_count_L_TO_R} / 실패 횟수 : {False_count_L_TO_R}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 모듈러 지수승 연산]\n')
+        p.write(f"[Multiplication and Squaring]\n실행 횟수 : {count} / 성공 횟수 : {count - False_count_MS} / 실패 횟수 : {False_count_MS}\n")
+        p.write(f"[Right to Left]\n실행 횟수 : {count} / 성공 횟수 : {count - False_count_R_TO_L} / 실패 횟수 : {False_count_R_TO_L}\n")
+        p.write(f"[Left to Right]\n실행 횟수 : {count} / 성공 횟수 : {count - False_count_L_TO_R} / 실패 횟수 : {False_count_L_TO_R}\n")
 
 def test_barrett(f, p):
     k = open('./result/result_barrett_reduction.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 barrett reduction 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -357,12 +367,13 @@ def test_barrett(f, p):
             k.write(shiftif)
             k.write(str(hex(tmp)))
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 barrett reduction 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def test_gcd(f, p):
     k = open('./result/result_gcd.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 gcd 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -383,7 +394,10 @@ def test_gcd(f, p):
             k.write(shiftif)
             k.write(str(hex(tmp)))
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 gcd 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def EEA(a, b):
     if b == 0:
@@ -396,8 +410,6 @@ def EEA(a, b):
 
 def test_EEA(f, p):
     k = open('./result/result_EEA.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 EEA 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -419,7 +431,10 @@ def test_EEA(f, p):
             k.write(str(" , "))
             k.write(str(hex(temp_y)))
             k.write('\n')
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 EEA 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def miller_rabin(n, k=10):
     if n < 2: return False
@@ -484,8 +499,6 @@ def is_prime_by_miller_rabin(n, k):
 def test_miller_rabin(f, p):
     k = open('./result/result_Miller_Rabin_fail.txt', 'w')
     t = open('./result/result_Miller_Rabin_success.txt', 'w')
-    p.write('------------------------------------------------------------\n')
-    p.write('[빅넘 Miller Rabin 연산]\n')
     count = 0
     False_count = 0
     while True:
@@ -499,7 +512,10 @@ def test_miller_rabin(f, p):
         # mr_result = is_prime_by_miller_rabin(value, 10)
         mr_result = miller_rabin(value, 10)
         False_count += write_result(k, t, shiftif, mr_result, result)
-    p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[빅넘 Miller Rabin 연산]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def write_result(k, t, shiftif, mr_result, result):
     if mr_result == False and 'Composite' in result:
@@ -514,6 +530,46 @@ def write_result(k, t, shiftif, mr_result, result):
     elif mr_result == True and 'Probably_Prime' in result:
         t.write(shiftif)
         return 0
+
+def test_RSA(f, p):
+    k = open('./result/result_RSA.txt', 'w')
+    count = 0
+    False_count = 0
+    while True:
+        shiftif = f.readline()
+        if not shiftif:
+            break
+        count += 1
+        shiftif_tmp = shiftif.split(' ')
+        result = shiftif_tmp[0]
+        if('RSA_FAIL' in result):
+            False_count += 1
+            k.write(shiftif)
+            k.write('\n')
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[RSA 수행 결과]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
+
+def test_RSA_CRT(f, p):
+    k = open('./result/result_RSA_CRT.txt', 'w')
+    count = 0
+    False_count = 0
+    while True:
+        shiftif = f.readline()
+        if not shiftif:
+            break
+        count += 1
+        shiftif_tmp = shiftif.split(' ')
+        result = shiftif_tmp[0]
+        if('RSA_FAIL' in result):
+            False_count += 1
+            k.write(shiftif)
+            k.write('\n')
+    if count != 0:
+        p.write('------------------------------------------------------------\n')
+        p.write('[RSA CRT 수행 결과]\n')
+        p.write(f"실행 횟수 : {count} / 성공 횟수 : {count - False_count} / 실패 횟수 : {False_count}\n")
 
 def bi_test(f, p):
     while True:
@@ -577,3 +633,9 @@ if __name__ == '__main__':
 
         f = file_open('./test_vec/test_miller_rabin.txt')
         if f != None:   test_miller_rabin(f, p)
+
+        f = file_open('./test_vec/test_RSA.txt')
+        if f != None: test_RSA(f, p)
+
+        f = file_open('./test_vec/test_RSA_CRT.txt')
+        if f != None: test_RSA_CRT(f, p)
