+++ test_init.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+# 디렉토리 생성
+mkdir -p ./ref/test/test_vec
+mkdir -p ./ref/test/result
+
+# ./ref/test 폴더로 이동
+cd ./ref/test
